requirements.txt
python-telegram-bot[ext]==20.8
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
beautifulsoup4==4.13.5
lxml==6.0.1